+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Fixed typo in `isConvenienceInitialiser` property
 - Fixed creating cache folder when cache is disabled
 - Fixed parsing multiple enum cases annotations
+- Fixed parsing inline annotations when there is an access level or attribute
 
 ### Internal changes